+++ src/RouterServiceProvider.php
@@ -10,8 +10,10 @@ class RouterServiceProvider extends ServiceProvider
 {
     /**
      * Bootstrap any application services.
+     *
+     * @return void
      */
-    public function boot()
+    public function boot() : void
     {
         $kernel = app(Kernel::class);